refactor: Update feature flag imports to use singular module name
